+++ package.json
@@ -2,14 +2,9 @@
   "name": "rate-my-professor-api-ts",
   "version": "0.0.216",
   "type": "module",
-  "main": "./dist/index.js",
-  "types": "./dist/index.d.ts",
-  "exports": {
-    ".": {
-      "types": "./dist/index.d.ts",
-      "import": "./dist/index.js"
-    }
-  },
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
+  "exports": "./dist/index.js",
   "files": [
     "dist"
   ],
